feat(live channel): split resource schema from resource structure and related expand/flatten functions EP-5205
