+++ aws/resource_aws_media_live_channel.go
@@ -950,20 +950,18 @@ func resourceAwsMediaLiveChannel() *schema.Resource {
 								Schema: map[string]*schema.Schema{
 									"source_end_behavior": {
 										Type:     schema.TypeString,
-										Required: true,
+										Optional: true,
 										Default:  "CONTINUE",
 									},
 
 									"input_filter": {
 										Type:     schema.TypeString,
 										Required: true,
-										Default:  "AUTO",
 									},
 
 									"filter_strength": {
 										Type:     schema.TypeInt,
 										Required: true,
-										Default:  1,
 									},
 
 									"deblock_filter": {
@@ -1094,7 +1092,7 @@ func resourceAwsMediaLiveChannelCreate(d *schema.ResourceData, meta interface{})
 
 	if v, ok := d.GetOk("encoder_settings"); ok && len(v.([]interface{})) > 0 {
 		input.EncoderSettings = expandEncoderSettings(
-			v.([]interface{}),
+			v.(*schema.Set),
 		)
 	}
 
@@ -1170,123 +1168,4 @@ func resourceAwsMediaLiveChannelUpdate(d *schema.ResourceData, meta interface{})
 
 func resourceAwsMediaLiveChannelDelete(d *schema.ResourceData, meta interface{}) error {
 	return nil
-}
-
-func expandInputAttachments(inputAttachments []interface{}) []*medialive.InputAttachment {
-	var result []*medialive.InputAttachment
-	if len(inputAttachments) == 0 {
-		return nil
-	}
-
-	for _, inputAtt := range inputAttachments {
-		r := inputAtt.(map[string]interface{})
-
-		result = append(result, &medialive.InputAttachment{
-			InputAttachmentName: aws.String(r["input_attachment_name"].(string)),
-			InputId:             aws.String(r["input_id"].(string)),
-			InputSettings:       expandInputAttachmentSettings(r["input_settings"]),
-		})
-	}
-	return result
-}
-
-func expandInputSpecification(s *schema.Set) *medialive.InputSpecification {
-	if s.Len() > 0 {
-		rawInputSpecification := s.List()[0].(map[string]interface{})
-		return &medialive.InputSpecification{
-			Codec:          aws.String(rawInputSpecification["codec"].(string)),
-			MaximumBitrate: aws.String(rawInputSpecification["maximum_bitrate"].(string)),
-			Resolution:     aws.String(rawInputSpecification["resolution"].(string)),
-		}
-	} else {
-		log.Printf("[WARN] MediaLive Channel: Input Specification can not be found")
-		return &medialive.InputSpecification{}
-	}
-}
-
-func expandDestinations(destinations []interface{}) []*medialive.OutputDestination {
-	var result []*medialive.OutputDestination
-	if len(destinations) == 0 {
-		log.Printf("[WARN] MediaLive Channel: At least one output destination required")
-		return nil
-	}
-
-	for _, destination := range destinations {
-		r := destination.(map[string]interface{})
-
-		result = append(result, &medialive.OutputDestination{
-			Id:       aws.String(r["id"].(string)),
-			Settings: expandOutputDestinationSettings(r["settings"].([]interface{})),
-		})
-	}
-	return result
-}
-
-func expandOutputDestinationSettings(destinationSettings []interface{}) []*medialive.OutputDestinationSettings {
-	var result []*medialive.OutputDestinationSettings
-	if len(destinationSettings) == 0 {
-		log.Printf("[ERROR] MediaLive Channel: One destination setting is required for each redundant encoder")
-		return nil
-	}
-
-	for _, settings := range destinationSettings {
-		r := settings.(map[string]interface{})
-
-		result = append(result, &medialive.OutputDestinationSettings{
-			PasswordParam: aws.String(r["password_param"].(string)),
-			StreamName:    aws.String(r["stream_name"].(string)),
-			Url:           aws.String(r["url"].(string)),
-			Username:      aws.String(r["username"].(string)),
-		})
-	}
-	return result
-}
-
-func expandEncoderSettings(s *schema.Set) *medialive.EncoderSettings {
-	if s.Len() > 0 {
-		rawEncoderSettings := s.List()[0].(map[string]interface{})
-		return &medialive.EncoderSettings{
-			AudioDescriptions: expandAudioDescriptions(rawEncoderSettings["audio_descriptions"].(string)),
-			OutputGroup:       expandOutputGroups(rawEncoderSettings["output_groups"]),
-			TimecodeConfig:    expandTimecodeConfigs(rawEncoderSettings["timecode_config"].(string)),
-			VideoDescriptions: expandVideoDescriptions(rawEncoderSettings["video_descriptions"].(string)),
-		}
-	} else {
-		log.Printf("[ERROR] MediaLive Channel: Encoder settings required")
-		return &medialive.InputSpecification{}
-	}
-}
-
-func expandAudioDescriptions(audioDescriptions []interface{}) []*medialive.AudioDescription {
-	var result []*medialive.AudioDescription
-	if len(audioDescriptions) == 0 {
-		log.Printf("[ERROR] MediaLive Channel: At least one audio description is required for each encoder")
-		return nil
-	}
-
-	for _, descs := range audioDescriptions {
-		r := descs.(map[string]interface{})
-
-		result = append(result, &medialive.OutputDestinationSettings{
-			AudioSelectorName: aws.String(r["audio_selector_name"].(string)),
-			Name:              aws.String(r["name"].(string)),
-			StreamName:        aws.String(r["stream_name"].(string)),
-			CodecSettings:     expandAudioCodecSettings(r["codec_settings"]),
-		})
-	}
-	return result
-}
-
-func expandAudioCodecSetting(s *schema.Set) *medialive.CodecSettings {
-	if s.Len() > 0 {
-		rawCodecSettings := s.List()[0].(map[string]interface{})
-		return &medialive.InputSpecification{
-			AacSettings:    expandAacCodecSettings(rawInputSpecification["aac_settings"].((*schema.Set))),
-			MaximumBitrate: aws.String(rawInputSpecification["maximum_bitrate"].(string)),
-			Resolution:     aws.String(rawInputSpecification["resolution"].(string)),
-		}
-	} else {
-		log.Printf("[WARN] MediaLive Channel: Input Specification can not be found")
-		return &medialive.InputSpecification{}
-	}
 }
